+++ src/FakeXrmEasy.Core/Extensions/XmlExtensionsForFetchXml.cs
@@ -7,6 +7,7 @@ using Microsoft.Xrm.Sdk;
 using System.Globalization;
 using FakeXrmEasy.Abstractions;
 using FakeXrmEasy.Abstractions.Exceptions;
+using FakeXrmEasy.Core.Exceptions.Query.FetchXml;
 
 namespace FakeXrmEasy.Extensions.FetchXml
 {
@@ -898,6 +899,26 @@ namespace FakeXrmEasy.Extensions.FetchXml
             return !OperatorsNotToConvertArray.Contains(conditionOperator);
         }
 
+        internal static object GetConditionExpressionBasedOnAttributeType(Type attributeType, string value, string sEntityName, string sAttributeName, ConditionOperator op)
+        {
+            try
+            {
+                if (ValueNeedsConverting(op))
+                {
+                    return GetValueBasedOnType(attributeType, value);
+                }
+
+                else
+                {
+                    return int.Parse(value);
+                }
+            }
+            catch (Exception e)
+            {
+                throw new Exception(string.Format("When trying to parse value for entity {0} and attribute {1}: {2}", sEntityName, sAttributeName, e.Message));
+            }
+        }
+        
         /// <summary>
         /// 
         /// </summary>
@@ -919,23 +940,16 @@ namespace FakeXrmEasy.Extensions.FetchXml
                     var attributeType = ctx.FindReflectedAttributeType(reflectedType, sEntityName, sAttributeName);
                     if (attributeType != null)
                     {
-                        try
-                        {
-                            if (ValueNeedsConverting(op))
-                            {
-                                return GetValueBasedOnType(attributeType, value);
-                            }
-
-                            else
-                            {
-                                return int.Parse(value);
-                            }
-                        }
-                        catch (Exception e)
-                        {
-                            throw new Exception(string.Format("When trying to parse value for entity {0} and attribute {1}: {2}", sEntityName, sAttributeName, e.Message));
-                        }
+                        return GetConditionExpressionBasedOnAttributeType(attributeType, value, sEntityName, sAttributeName, op);
                     }
+                }
+            }
+            else
+            {
+                var injectedAttributeType = ctx.FindAttributeTypeInInjectedMetadata(sEntityName, sAttributeName);
+                if (injectedAttributeType != null)
+                {
+                    return GetConditionExpressionBasedOnAttributeType(injectedAttributeType, value, sEntityName, sAttributeName, op);
                 }
             }
 
@@ -976,7 +990,7 @@ namespace FakeXrmEasy.Extensions.FetchXml
 
             if (bIsNumeric || bIsDateTime)
             {
-                throw new Exception($"When using arithmetic values in a condition of attribute '{sAttributeName}' of entity '{sEntityName}' in a Fetch a ProxyTypesAssembly must be used in order to know which types to cast values to. If you are using early bound types, please make sure the early bound type was generated for entity '{sEntityName}'");
+                throw new ArithmeticTypeConversionException(sEntityName, sAttributeName);
             }
 
             //Default value