+++ src/FakeXrmEasy.Core/FileStorage/Db/FileBlock.cs
@@ -1,0 +1,8 @@
+namespace FakeXrmEasy.Core.FileStorage.Db
+{
+    internal class FileBlock
+    {
+        internal string BlockId { get; set; }
+        internal byte[] Content { get; set; }
+    }
+}