+++ src/FakeXrmEasy.Core/FileStorage/Db/FileDownloadSession.cs
@@ -1,0 +1,12 @@
+using FakeXrmEasy.Abstractions.FileStorage;
+using FakeXrmEasy.Core.FileStorage.Download;
+
+namespace FakeXrmEasy.Core.FileStorage.Db
+{
+    internal class FileDownloadSession
+    {
+        internal string FileDownloadSessionId { get; set; }
+        internal FileDownloadProperties Properties { get; set; }
+        internal IFileAttachment File { get; set; }
+    }
+}