+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [3.6.0]
+
+### Added
+
+- Added new InMemoryFileDb implementation to support file and image storage - https://github.com/DynamicsValue/fake-xrm-easy/issues/157
+- Added default max file size for file and image uploads - https://github.com/DynamicsValue/fake-xrm-easy/issues/157
+
+### Changed
+
+- Resolves many issues in Create, Upsert, and CreateMultiple and UpsertMultiple when alternate keys were used - https://github.com/DynamicsValue/fake-xrm-easy/issues/172 
+- TracingService will also output to the default standard output - https://github.com/DynamicsValue/fake-xrm-easy/issues/163
+- **BREAKING CHANGE**: This will **only** affect you if you use XrmRealContext class. Moved XrmRealContext to a separate FakeXrmEasy.Integration package to remove dependency on XrmTooling - https://github.com/DynamicsValue/fake-xrm-easy/issues/160
+- Resolves an issue in FetchXml queries when using arithmetic values and no early bound assemblies are used. It will now read from injected metadata in absence of proxy type assemblies - https://github.com/DynamicsValue/fake-xrm-easy/issues/158 
+- Resolves issue in MetadataGenerator where relationship properties were generated in the wrong order, also generates ManyToMany relationship properties - https://github.com/DynamicsValue/fake-xrm-easy/issues/135
+- Adds implementation of RelatedEntities in Update message , before it was implemented only for Create - https://github.com/DynamicsValue/fake-xrm-easy/issues/154
 
 ## [3.5.1]
 
@@ -9,7 +24,7 @@
 
 ### Added
 
-- Added FileAttributeMetadata support to MetadataGenerator 
+- Added FileAttributeMetadata support to MetadataGenerator
 - Added support for bulk operations: CreateMultipleRequest, UpdateMultipleRequest, UpsertMultipleRequest - https://github.com/DynamicsValue/fake-xrm-easy/issues/122
 - Added new exception to make the initialization of entity records with attributes with a null entity reference more obvious (thanks Betim) - https://github.com/DynamicsValue/fake-xrm-easy/issues/107
 - Add support for OptionSetValueCollection attributes when they are generated as an IEnumerable<TEnum> (using EBG or pac modelbuilder) - https://github.com/DynamicsValue/fake-xrm-easy/issues/140
@@ -17,12 +32,12 @@
 
 ### Changed
 
-- Improves exception message when an early bound type was not generated - https://github.com/DynamicsValue/fake-xrm-easy/issues/145 
+- Improves exception message when an early bound type was not generated - https://github.com/DynamicsValue/fake-xrm-easy/issues/145
 - Resolves referencing EntityAlias or EntityName in conditions inside nested filters of a LinkedEntity (thanks Temmy) - https://github.com/DynamicsValue/fake-xrm-easy/issues/63
 - Resolves Resolving entity references by Alternate Keys when EntityMetadata doesn't have any Keys. - https://github.com/DynamicsValue/fake-xrm-easy/issues/138
 - Resolves an issue where a ConditionExpression with an In operator should to not take array of integers as an input, but instead separate values (thanks Ben and Betim) - https://github.com/DynamicsValue/fake-xrm-easy/issues/96
 - Resolves filtering Money attributes by an integer value (thanks Ben and Betim) - https://github.com/DynamicsValue/fake-xrm-easy/issues/64
-- Upgraded DataverseClient dependency to 1.1.22 
+- Upgraded DataverseClient dependency to 1.1.22
 
 
 ## [3.4.2]
@@ -62,10 +77,10 @@
 ### Changed
 
 -  Update namespaces in tests project for consistency
- - Upgraded GitHub Actions to update Java major version to run SonarCloud analysis - https://github.com/DynamicsValue/fake-xrm-easy/issues/110
- - Introduced new NewEntityRecord method to easily create instances of entity records based on the current use of early-bound or late-bound entities 
- - Resolves an issue with query evaluation and MultiOptionSets when using late bound entities or if type information is not present. - https://github.com/DynamicsValue/fake-xrm-easy/issues/66
- - Update legacy CRM SDK 2011 dependency to use official MS package - https://github.com/DynamicsValue/fake-xrm-easy/issues/105
+- Upgraded GitHub Actions to update Java major version to run SonarCloud analysis - https://github.com/DynamicsValue/fake-xrm-easy/issues/110
+- Introduced new NewEntityRecord method to easily create instances of entity records based on the current use of early-bound or late-bound entities
+- Resolves an issue with query evaluation and MultiOptionSets when using late bound entities or if type information is not present. - https://github.com/DynamicsValue/fake-xrm-easy/issues/66
+- Update legacy CRM SDK 2011 dependency to use official MS package - https://github.com/DynamicsValue/fake-xrm-easy/issues/105
 
 ## [3.3.1]
 
@@ -92,6 +107,7 @@
 ## [3.1.2]
 
 - Bump dataverse dependency to production ready version 1.0.1
+
 ## [3.1.1]
 
 ### Changed
@@ -112,7 +128,7 @@
 
 ## [3.0.2]
 
-### Changed 
+### Changed
 
 - Bump Dataverse dependency to 0.6.1 from 0.5.10 to solve DynamicsValue/fake-xrm-easy#20
 - Also replaced Microsoft.Dynamics.Sdk.Messages dependency, as it has also been deprecated by MSFT, to Microsoft.PowerPlatform.Dataverse.Client.Dynamics 0.6.1 DynamicsValue/fake-xrm-easy#20